Publish faces in different frame than CameraTop
This is important because we still publish the transform to a face even
if the face is not visible anymore. If the transform is bound to the head
the human position will be useless as soon as the robot moves the head.

So it may be usefull to publish the human pose in a 'fixed' frame like odom.

It does not work well yet because of TF transformation issues, related to
timestamps.